refactor(daemon): split daemon.py into server + handlers (#59)
Move worker loop + `_Job` + three handlers (`_handle_blend`, `_handle_encode`,
`_handle_job`) into `src/imagecli/daemon_handlers.py`. Keep server loop, wire
protocol, public client API, and startup model loaders in `daemon.py`.

- `daemon.py`: 225 LOC (was 459)
- `daemon_handlers.py`: 240 LOC (new)
- Remove `daemon.py` exemption from `tools/file_exemptions.txt`

Loaders (`load_pipe`, `load_encoder`, `_check_vram`) stay in `daemon.py`
because they are called once at startup from `daemon_main` and keeping
them here leaves the handlers module below the 300 LOC gate without
splitting a third file.

Wire format and handler signatures are bit-for-bit unchanged; only
call sites and module boundaries moved. Public API
(`daemon_main`, `daemon_request`, `SOCKET_PATH`) unchanged.

Closes #59
